+++ Cargo.toml
@@ -3,7 +3,7 @@ members = [
   "bin/pcl", 
   "crates/phoundry",
   "crates/common",
-  "crates/da",
+  "crates/core"
 ]
 default-members = ["bin/pcl"]
 resolver = '2'
@@ -22,7 +22,7 @@ repository = "https://github.com/phylaxsystems/pcl"
 [workspace.dependencies]
 pcl-phoundry = { path = "crates/phoundry" }
 pcl-common = { path = "crates/common" }
-pcl-da = { path = "crates/da" }
+pcl-core= { path = "crates/core" }
 tokio = { version = "1.39.0", features = ["full"] }
 clap = { version = "4.5.23", features = ["derive", "env"] }
 eyre = "0.6.12"
